requirements.txt
asgiref==3.9.1
Django==5.2.5
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.1
et-xmlfile==1.1.0
openpyxl==3.1.5
pillow==11.3.0
sqlparse==0.5.3
typing_extensions==4.14.1
